Lensesdata
Original Source:

Cendrowska, J. "PRISM: An algorithm for inducing modular rules", International Journal of Man-Machine Studies, 1987, 27, 349-370

Donor:

Benoit Julien (Julien '@' ce.cmu.edu)
Data Set Information:

The examples are complete and noise free. The examples highly simplified the problem. The attributes do not fully describe all the factors affecting the decision as to which type, if any, to fit.

Notes:

--This database is complete (all possible combinations of attribute-value pairs are represented).

--Each instance is complete and correct.

--9 rules cover the training set.


Attribute Information:

3 Classes
the patient should be fitted with hard contact lenses,
the patient should be fitted with soft contact lenses,
the patient should not be fitted with contact lenses.

1. age of the patient: (1) young, (2) pre-presbyopic, (3) presbyopic
2. spectacle prescription: (1) myope, (2) hypermetrope
3. astigmatic: (1) no, (2) yes
4. tear production rate: (1) reduced, (2) normal